+++ spotbugs-tests/src/test/java/edu/umd/cs/findbugs/detect/Issue2793Test.java
@@ -18,11 +18,15 @@ class Issue2793Test extends AbstractIntegrationTest {
     void testRecordSerialVersionUid() {
         performAnalysis("../java17/ghIssues/Issue2793.class",
                 "../java17/ghIssues/Issue2793$RecordWithoutSerialVersionUid.class",
-                "../java17/ghIssues/Issue2793$ExternalizableRecord.class");
+                "../java17/ghIssues/Issue2793$ExternalizableRecord.class",
+                "../java17/ghIssues/Issue2793$SerializableClass.class",
+                "../java17/ghIssues/Issue2793$YangLibModule.class");
 
-        assertBugCount("SE_NO_SERIALVERSIONID", 0);
+        assertBugCount("SE_NO_SERIALVERSIONID", 2);
 
         assertBugCount("SE_NO_SUITABLE_CONSTRUCTOR_FOR_EXTERNALIZATION", 0);
+
+        assertBugCount("SE_BAD_FIELD", 0);
     }
 
     private void assertBugCount(String type, int expectedCount) {